refactor: move number-field core and base CSS to styles folder
