+++ packages/number-field/src/styles/vaadin-number-field-base-styles.js
