+++ packages/number-field/src/styles/vaadin-number-field-core-styles.js
